Create README.md
I have implemented a wall clock in VHDL. Its format is in XX:XX:XX form, 23:59:45 for example. I have used 6 7-segment displays for the clock, and 3 buttons one is for reset, one is for entering and leaving time setting mode, and the other one is for in time setting mode selecting one of the six numbers to adjust. While in time setting mode, the led under the number to be adjusted is blinking in order to indicate which number to be adjusted. This project is not tested on an FPGA yet.